[feature] 敵侵攻のベース実装
