+++ Assets/Prefabs/Camera.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 9fc0d4010bbf28b4594072e72b8655ab
-DefaultImporter:
+guid: e4d852e93b0812b43a16cb9ed92aae64
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
