+++ Assets/Scripts/AClass/AEnemy.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 5ed4092d1bd64aa8897a00da060812ee
+timeCreated: 1720716023
