+++ Assets/Scripts/Enemies/TestEnemy.cs
@@ -1,0 +1,8 @@
+using AClass;
+
+namespace Enemies
+{
+    public class TestEnemy : AEnemy
+    {
+    }
+}
