+++ Assets/Scripts/InvasionPhase/InvasionMazeController.cs
@@ -1,3 +1,4 @@
+using AClass;
 using DataClass;
 using UnityEngine;
 using UnityEngine.Serialization;
