+++ Assets/Scripts/ScriptableObjects/InvationData.cs
@@ -1,0 +1,6 @@
+namespace ScriptableObjects
+{
+    public class InvasionData
+    {
+    }
+}
