+++ Assets/Scripts/ScriptableObjects/InvationData.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7e6d924fc56f480f86369736414b4688
+timeCreated: 1720713664
