+++ Assets/Scripts/Traps/TestTrap.cs
@@ -1,3 +1,4 @@
+using AClass;
 using UnityEngine;
 
 namespace Traps
